move all to same file, and only raise error when version too small
